Create server method for getToolResponse function and use it in schema and path routes
